Directory & Services Updated

+++ database/migrations/2021_03_24_000024_add_relationship_fields_to_services_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToServicesTable extends Migration
     public function up()
     {
         Schema::table('services', function (Blueprint $table) {
-            $table->unsignedInteger('servicessubcategory_id')->nullable();
+            $table->unsignedBigInteger('servicessubcategory_id')->nullable();
             $table->foreign('servicessubcategory_id', 'servicessubcategory_fk_2384531')->references('id')->on('services_sub_categories');
         });
     }
